timescale-agents/pyproject.toml
[tool.poetry]
name = "timescale-agents"
version = "0.1.0"
description = ""
authors = ["Jolkkonen, Johannes (External) <johannes.jolkkonen@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
shiny = "^0.10.2"
shinywidgets = "^0.3.2"
plotly = "^5.22.0"
pandas = "^2.2.2"
fastapi = "^0.111.0"
apscheduler = "^3.10.4"
psycopg2 = "^2.9.9"
python-decouple = "^3.8"
requests = "^2.32.3"
langchain = "^0.2.5"
langchain-openai = "^0.1.8"
tenacity = "^8.4.1"
tabulate = "^0.9.0"
black = "^24.4.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
